requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
matplotlib>=3.7.0
segment-anything>=1.0
clip @ git+https://github.com/openai/CLIP.git
scikit-image>=0.20.0
openai>=1.3.0
stability-sdk>=0.8.0
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.65.0
safetensors>=0.3.0 